WIP: add CommentEditor component
